+++ activities/views/response_invited.php
@@ -11,5 +11,5 @@ $formatter = new CalendarDateFormatter(['calendarItem' => $source]);
 
 echo Yii::t('CalendarModule.views', '%displayName% is invited to %contentTitle%.', [
     '%displayName%' => '<strong>' . Html::encode($originator->displayName) . '</strong>',
-    '%contentTitle%' => $this->context->getContentInfo($source).' on '.$formatter->getFormattedTime()
+    '%contentTitle%' => Html::encode($this->context->getContentInfo($source)).' on '.$formatter->getFormattedTime()
 ]);