+++ messages/nl/export.php
@@ -1,14 +1,13 @@
 <?php
-
 return [
-    '<strong>Calendar</strong> export' => '',
-    'CalDAV Base URL' => '',
-    'CalDAV URL' => '',
-    'Copy to clipboard' => '',
-    'Password' => '',
-    'This CalDAV link lets you sync events both ways with your calendar app. It supports two-way sync-changes made in your calendar app will be reflected in HumHub, and vice versa. Some calendar apps require the base URL, while others only need the full CalDAV URL. Use the credentials from your HumHub account when prompted.' => '',
-    'This link can be added to your calendar for syncing events. iCal updates your app with new events but doesn’t sync changes back (one-way sync).' => '',
-    'Username' => '',
-    'Your HumHub password' => '',
-    'iCal URL' => '',
+  '<strong>Calendar</strong> export' => '<strong>Agenda</strong> exporteren',
+  'CalDAV Base URL' => 'CalDAV-basis-URL',
+  'CalDAV URL' => 'CalDAV-URL',
+  'Copy to clipboard' => 'Kopiëren naar klembord',
+  'Password' => 'Wachtwoord',
+  'This CalDAV link lets you sync events both ways with your calendar app. It supports two-way sync-changes made in your calendar app will be reflected in HumHub, and vice versa. Some calendar apps require the base URL, while others only need the full CalDAV URL. Use the credentials from your HumHub account when prompted.' => 'Met deze CalDAV-koppeling kunt u afspraken in beide richtingen synchroniseren met uw agenda-app. De koppeling ondersteunt synchronisatie in beide richtingen: wijzigingen in je agenda-app worden doorgevoerd in HumHub, en vice versa. Sommige agenda-apps vereisen de basis-URL, terwijl andere alleen de volledige CalDAV-URL nodig hebben. Gebruik de inloggegevens van je HumHub-account wanneer daarom wordt gevraagd.',
+  'This link can be added to your calendar for syncing events. iCal updates your app with new events but doesn’t sync changes back (one-way sync).' => 'Deze link kan aan uw agenda worden toegevoegd om afspraken te synchroniseren. iCal werkt uw app bij met nieuwe afspraken, maar synchroniseert geen wijzigingen terug (eenrichtingssynchronisatie).',
+  'Username' => 'Gebruikersnaam',
+  'Your HumHub password' => 'Uw HumHub-wachtwoord',
+  'iCal URL' => 'iCal-URL',
 ];