+++ messages/pt-BR/export.php
@@ -1,14 +1,13 @@
 <?php
-
 return [
-    '<strong>Calendar</strong> export' => '',
-    'CalDAV Base URL' => '',
-    'CalDAV URL' => '',
-    'Copy to clipboard' => '',
-    'Password' => '',
-    'This CalDAV link lets you sync events both ways with your calendar app. It supports two-way sync-changes made in your calendar app will be reflected in HumHub, and vice versa. Some calendar apps require the base URL, while others only need the full CalDAV URL. Use the credentials from your HumHub account when prompted.' => '',
-    'This link can be added to your calendar for syncing events. iCal updates your app with new events but doesn’t sync changes back (one-way sync).' => '',
-    'Username' => '',
-    'Your HumHub password' => '',
-    'iCal URL' => '',
+  '<strong>Calendar</strong> export' => 'Exportação de <strong>calendário</strong>',
+  'CalDAV Base URL' => 'URL base do CalDAV',
+  'CalDAV URL' => 'URL CalDAV',
+  'Copy to clipboard' => 'Copiar para a área de transferência',
+  'Password' => 'Senha',
+  'This CalDAV link lets you sync events both ways with your calendar app. It supports two-way sync-changes made in your calendar app will be reflected in HumHub, and vice versa. Some calendar apps require the base URL, while others only need the full CalDAV URL. Use the credentials from your HumHub account when prompted.' => 'Este link do CalDAV permite sincronizar eventos nos dois sentidos com o seu aplicativo de calendário. Ele suporta sincronização bidirecional: as alterações feitas no seu aplicativo de calendário serão refletidas no HumHub e vice-versa. Alguns aplicativos de calendário exigem a URL base, enquanto outros exigem apenas a URL completa do CalDAV. Use as credenciais da sua conta do HumHub quando solicitado.',
+  'This link can be added to your calendar for syncing events. iCal updates your app with new events but doesn’t sync changes back (one-way sync).' => 'Este link pode ser adicionado ao seu calendário para sincronizar eventos. O iCal atualiza seu aplicativo com novos eventos, mas não sincroniza as alterações de volta (sincronização unidirecional).',
+  'Username' => 'Nome de Usuário',
+  'Your HumHub password' => 'Sua senha do HumHub',
+  'iCal URL' => 'URL iCal',
 ];