+++ docs/CHANGELOG.md
@@ -3,7 +3,8 @@ Changelog
 
 1.7.8 (Unreleased)
 ---------------------
-- Enh #557: Use content ID in the "Custom Pages" extension
+- Enh #557: Use content ID in the `Custom Pages` extension
+- Enh #561: Add calendar name for `ical`
 
 1.7.7 (June 30, 2025)
 ---------------------