+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog
 ---------------------
 - Enh #557: Use content ID in the "Custom Pages" extension
 - Enh #562: Add JWT token for caldav password
+- Enh #557: Use content ID in the `Custom Pages` extension
+- Enh #561: Add calendar name for `ical`
 
 1.7.7 (June 30, 2025)
 ---------------------